test(product): add unit tests for createProduct endpoint - cover successful creation - cover invalid argument (validation failure) returning 400 Bad Request

+++ backend/src/main/java/com/merchordersystem/backend/dto/product/ProductRequest.java
@@ -13,10 +13,10 @@ public class ProductRequest {
     private String name;
 
     @NotNull
-    private double price;
+    private Double price;
 
     @NotNull
-    private double number;
+    private Integer number;
 
     @NotNull
     private String description;
